refactor: move reverse geocode into utils folder
